Adding backened and frontend component for scatter plot
Still need to add visualization page as well as test to see what the graph looks like and view the information it displays. Will need adjustments and add more to this code once reviewed as its currently a basic structure.  Will also need to make adjustments to work with our cleaned data,